Разработка http://jira.aplana.com/browse/SBRFACCTAX-3417 : Вынес импорт в виджет, добавил вывод ошибок в лог.

+++ core/src/main/java/com/aplana/sbrf/taxaccounting/service/FormDataService.java
@@ -55,7 +55,7 @@ public interface FormDataService {
      * @param inputStream
      * @param fileName
      */
-    public void importFormData(Logger logger, TAUserInfo userInfo, int formDataId,int formTemplateId,
+    public void importFormData(Logger logger, TAUserInfo userInfo, long formDataId,int formTemplateId,
                                int departmentId, FormDataKind kind, int reportPeriodId, InputStream inputStream, String fileName);
